Move CiSkippedTheory to common proj
